1125 move switchSearchMarkup, initPaginationListeners & scrollToTopOfSearch helper functions to _helpers.js file
1125 create _helpers.js file
